requirements.txt
ago==0.0.93
attrs==21.4.0
Automat==20.2.0
awscli==1.22.37
beautifulsoup4==4.10.0
bert-extractive-summarizer==0.10.1
blis==0.7.5
botocore==1.23.37
bs4==0.0.1
catalogue==2.0.6
cchardet==2.1.7
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.10
click==8.0.3
colorama==0.4.3
commonmark==0.9.1
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
cymem==2.0.6
docutils==0.15.2
dotmap==1.3.26
elasticsearch==7.16.3
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.4.2
h2==3.2.0
hjson==3.0.2
hpack==3.0.0
huggingface-hub==0.4.0
hurry.filesize==0.9
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jieba3k==0.35.1
Jinja2==3.0.3
jmespath==0.10.0
joblib==1.1.0
keybert==0.5.0
langcodes==3.3.0
langdetect==1.0.9
lxml==4.7.1
MarkupSafe==2.0.1
murmurhash==1.0.6
news-please==1.5.21
newspaper3k==0.2.8
nltk==3.6.7
numpy==1.22.1
packaging==21.3
parsel==1.6.0
pathy==0.6.1
Pillow==9.0.0
plac==1.3.4
preshed==3.0.6
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.8.2
PyDispatcher==2.0.5
Pygments==2.11.2
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-dateutil==2.8.2
pywin32==303
PyYAML==5.4.1
queuelib==1.6.2
readability-lxml==0.8.1
regex==2021.11.10
requests==2.27.1
requests-file==1.5.1
rich==11.0.0
rsa==4.7.2
s3transfer==0.5.0
sacremoses==0.0.47
scikit-learn==1.0.2
scipy==1.7.3
Scrapy==2.5.1
sentence-transformers==2.1.0
sentencepiece==0.1.96
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==5.2.1
soupsieve==2.3.1
spacy==3.2.1
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
thinc==8.0.13
threadpoolctl==3.0.0
tinysegmenter==0.3
tldextract==3.1.2
tokenizers==0.10.3
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
transformers==4.15.0
tweepy==4.5.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
typer==0.4.0
typing_extensions==4.0.1
urllib3==1.26.8
w3lib==1.22.0
warcio==1.7.4
wasabi==0.9.0
zope.interface==5.4.0
